requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
djangorestframework==3.11.1
drf-yasg==1.17.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
iniconfig==1.0.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.9.0
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
